+++ CookIT/CookIT/CookIT.Model/Factories/IngredientFactory.cs
@@ -9,7 +9,7 @@ namespace CookIT.Model.Factories
     public class IngredientFactory
     {
         //(int inID, string inName, int energy, int carbs, int proteins, int fats, int fibs, float salt, float minerals)
-        public static Ingredient CreateIngredient(int inID, string name, int energy, int carbs, int proteins, int fats, int fibs, float salt, float minerals)
+        public static Ingredient CreateIngredient(int inID, string name, float energy, float carbs, float proteins, float fats, float fibs, float salt, float minerals)
         {
             Ingredient ingred = null;
             ingred = new Ingredient(inID, name, energy, carbs, proteins, fats, fibs, salt, minerals);
